xrCore: strictly-typed X-Ray pt I
* xrCore is now free of -Wconversion -Wnarrowing -Wold-style-cast;
* most of constructors in xrCore and lots in other places are now
  explicit. This allowed to optimize different pieces of code
  throughout the engine, notably the removal of an implicit
  czstring <-> shared_str conversion showed tons of places where
  the shared string container was used without reason on was emitting
  additional copying etc.
* all timestamps and file sizes are now of s64, no more 32 bits. Except
  for XDB which has u32s in its headers, so no chance for 64-bit file
  support;
* partial fix for trivial types like vectors, but with custom copy/move
  constructors to use 64-bit or 128-bit writes. [[clang::trivial_abi]]
  helped, but currently doesn't work well with the Windows/MSVC ABI, so
  more can be done in future;
* IReader now has its data pointer const, which uncovered lots of places
  to optimize (several class methods were reading directly from IReaders,
  so making them return const pointers helped);
* XR_SYSV is dropped -- since it can't be used for constructors, the
  whole idea of making the whole engine use the SysV ABI is not feasible
  currently; and marking only some functions brings more harm than good
  (calling a SysV function from an MS function costs a dozen of
  instructions due to the XMM clobbering).
  If/when Clang is able to apply the attribute to constructors and
  destructors and ideally using a pragma instead of a per-function
  attribute, I might go back try this again;
* lots of optimization, constexpr, hardening, etc.

+++ ogsr_engine/xrCore/_compressed_normal.cpp
@@ -20,7 +20,7 @@ namespace
 class pvUVAdjustment_init
 {
 private:
-    std::array<float, 0x2000> tbl;
+    std::array<f32, 0x2000> tbl;
 
 public:
     constexpr pvUVAdjustment_init()
@@ -38,16 +38,16 @@ public:
             }
 
             // convert to 3D vectors
-            float x{gsl::narrow<float>(xbits)};
-            float y{gsl::narrow<float>(ybits)};
-            float z{gsl::narrow<float>(126 - xbits - ybits)};
+            f32 x{gsl::narrow<f32>(xbits)};
+            f32 y{gsl::narrow<f32>(ybits)};
+            f32 z{gsl::narrow<f32>(126 - xbits - ybits)};
 
             // calculate the amount of normalization required
             elem = 1.0f / _sqrt(y * y + z * z + x * x);
         }
     }
 
-    [[nodiscard]] constexpr const float& operator[](gsl::index i) const { return tbl[i]; }
+    [[nodiscard]] constexpr const f32& operator[](gsl::index i) const { return tbl[gsl::narrow_cast<size_t>(i)]; }
 };
 
 const pvUVAdjustment_init pvUVAdjustment;
